Finish popover
Fixes #14

There is now a popup available to select Begins With or Contains for
first and last name as well as a popup for State Only on home address.
Setting Home Address to State Only changes the home address input view
to a picker containing all the states (plus DC).